+++ crates/miden-testing/src/kernel_tests/block/proposed_block_errors.rs
@@ -1,25 +1,33 @@
-use std::{collections::BTreeMap, vec::Vec};
+use core::slice;
+use std::collections::BTreeMap;
+use std::vec::Vec;
 
 use assert_matches::assert_matches;
-use miden_objects::{
-    MAX_BATCHES_PER_BLOCK, ProposedBlockError,
-    account::AccountId,
-    block::{BlockInputs, BlockNumber, ProposedBlock},
-    crypto::merkle::SparseMerklePath,
-    note::NoteInclusionProof,
-    testing::account_id::ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET,
-    transaction::{OutputNote, ProvenTransaction},
-};
+use miden_objects::account::AccountId;
+use miden_objects::block::{BlockInputs, BlockNumber, ProposedBlock};
+use miden_objects::crypto::merkle::SparseMerklePath;
+use miden_objects::note::NoteInclusionProof;
+use miden_objects::testing::account_id::ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET;
+use miden_objects::transaction::{OutputNote, ProvenTransaction};
+use miden_objects::{MAX_BATCHES_PER_BLOCK, ProposedBlockError};
 use vm_processor::crypto::MerklePath;
 
 use super::utils::{
-    TestSetup, generate_batch, generate_executed_tx_with_authenticated_notes,
-    generate_fungible_asset, generate_output_note, generate_tracked_note,
-    generate_tracked_note_with_asset, generate_tx_with_authenticated_notes,
-    generate_tx_with_expiration, generate_tx_with_unauthenticated_notes, generate_untracked_note,
+    TestSetup,
+    generate_batch,
+    generate_executed_tx_with_authenticated_notes,
+    generate_fungible_asset,
+    generate_output_note,
+    generate_tracked_note,
+    generate_tracked_note_with_asset,
+    generate_tx_with_authenticated_notes,
+    generate_tx_with_expiration,
+    generate_tx_with_unauthenticated_notes,
+    generate_untracked_note,
     setup_chain,
 };
-use crate::{ProvenTransactionExt, utils::create_spawn_note};
+use crate::ProvenTransactionExt;
+use crate::utils::create_spawn_note;
 
 /// Tests that too many batches produce an error.
 #[test]
@@ -315,7 +323,8 @@ fn proposed_block_fails_on_invalid_proof_or_missing_note_inclusion_reference_blo
     let note0 = generate_untracked_note(account0.id(), account1.id());
 
     // This tx will use block1 as the reference block.
-    let tx0 = generate_tx_with_unauthenticated_notes(&mut chain, account1.id(), &[note0.clone()]);
+    let tx0 =
+        generate_tx_with_unauthenticated_notes(&mut chain, account1.id(), slice::from_ref(&note0));
 
     // This batch will use block1 as the reference block.
     let batch0 = generate_batch(&mut chain, vec![tx0]);
@@ -389,7 +398,8 @@ fn proposed_block_fails_on_missing_note_inclusion_proof() -> anyhow::Result<()> 
 
     let note0 = generate_tracked_note(&mut chain, account0.id(), account1.id());
 
-    let tx0 = generate_tx_with_unauthenticated_notes(&mut chain, account1.id(), &[note0.clone()]);
+    let tx0 =
+        generate_tx_with_unauthenticated_notes(&mut chain, account1.id(), slice::from_ref(&note0));
 
     let batch0 = generate_batch(&mut chain, vec![tx0]);
 
@@ -416,7 +426,8 @@ fn proposed_block_fails_on_missing_nullifier_witness() -> anyhow::Result<()> {
     let note0 = generate_untracked_note(account0.id(), account1.id());
 
     // This tx will use block1 as the reference block.
-    let tx0 = generate_tx_with_unauthenticated_notes(&mut chain, account1.id(), &[note0.clone()]);
+    let tx0 =
+        generate_tx_with_unauthenticated_notes(&mut chain, account1.id(), slice::from_ref(&note0));
 
     // This batch will use block1 as the reference block.
     let batch0 = generate_batch(&mut chain, vec![tx0]);
@@ -454,7 +465,8 @@ fn proposed_block_fails_on_spent_nullifier_witness() -> anyhow::Result<()> {
     let note0 = generate_untracked_note(account0.id(), account1.id());
 
     // This tx will use block1 as the reference block.
-    let tx0 = generate_tx_with_unauthenticated_notes(&mut chain, account1.id(), &[note0.clone()]);
+    let tx0 =
+        generate_tx_with_unauthenticated_notes(&mut chain, account1.id(), slice::from_ref(&note0));
 
     // This batch will use block1 as the reference block.
     let batch0 = generate_batch(&mut chain, vec![tx0]);