+++ crates/miden-testing/src/kernel_tests/tx/test_input_note.rs
@@ -1,7 +1,8 @@
 use alloc::string::String;
 
-use miden_lib::transaction::TransactionKernel;
-use miden_objects::{Word, note::Note, transaction::TransactionScript};
+use miden_lib::utils::ScriptBuilder;
+use miden_objects::Word;
+use miden_objects::note::Note;
 
 use super::{TestSetup, setup_test, word_to_masm_push_string};
 use crate::TxContextInput;
@@ -76,7 +77,7 @@ fn test_get_asset_info() -> anyhow::Result<()> {
         ),
     );
 
-    let tx_script = TransactionScript::compile(code, TransactionKernel::testing_assembler())?;
+    let tx_script = ScriptBuilder::with_kernel_library()?.compile_tx_script(code)?;
 
     let tx_context = mock_chain
         .build_tx_context(
@@ -134,7 +135,7 @@ fn test_get_recipient_and_metadata() -> anyhow::Result<()> {
         METADATA = word_to_masm_push_string(&p2id_note_1_asset.metadata().into()),
     );
 
-    let tx_script = TransactionScript::compile(code, TransactionKernel::testing_assembler())?;
+    let tx_script = ScriptBuilder::with_kernel_library()?.compile_tx_script(code)?;
 
     let tx_context = mock_chain
         .build_tx_context(TxContextInput::AccountId(account.id()), &[], &[p2id_note_1_asset])?
@@ -225,7 +226,7 @@ fn test_get_assets() -> anyhow::Result<()> {
         check_note_2 = check_assets_code(2, 8, &p2id_note_2_assets),
     );
 
-    let tx_script = TransactionScript::compile(code, TransactionKernel::testing_assembler())?;
+    let tx_script = ScriptBuilder::with_kernel_library()?.compile_tx_script(code)?;
 
     let tx_context = mock_chain
         .build_tx_context(